+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- aly broke MoveTo, this fixes it
 - Config: Preferred crafting/gathering job by alydev
 - Quests: automatic switching to configured job if current job is invalid for quest by alydev
   - ^ only on AcceptQuest for now by alydev